Move exitcode under exit
